requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-filter==24.3
pdfplumber==0.11.2
Pillow==11.1.0
gunicorn==23.0.0
